Move more types to cdf client
