+++ src/cdf/types.ts
@@ -1,0 +1,145 @@
+
+export type Datapoints = Items<Datapoint>;
+export type Events = Items<Event>;
+
+export interface Datapoint {
+  id: number;
+  externalId?: string;
+  isStep: boolean;
+  isString: boolean;
+  unit?: string;
+  datapoints: TimeSeriesDatapoint[] | TimeSeriesAggregateDatapoint[];
+}
+
+export interface TimeSeriesDatapoint {
+  timestamp: number;
+  value: string;
+}
+
+export interface TimeSeriesAggregateDatapoint {
+  timestamp: number;
+  average?: number;
+  max?: number;
+  min?: number;
+  count?: number;
+  sum?: number;
+  interpolation?: number;
+  stepInterpolation?: number;
+  continuousVariance?: number;
+  discreteVariance?: number;
+  totalVariation?: number;
+}
+
+export interface Event {
+  id: number;
+  startTime: number;
+  endTime: number;
+  description: string;
+  type: string;
+  subtype: string;
+  assetIds: number[];
+  source: string;
+  sourceId: string;
+}
+
+export interface FilterRequestParams {
+  metadata?: Metadata;
+  assetSubtreeIds?: IdEither[];
+  createdTime?: TimeRange;
+  lastUpdatedTime?: TimeRange;
+  externalIdPrefix?: string;
+}
+
+export interface AssetsFilterRequestParams extends FilterRequestParams {
+  name?: string;
+  parentIds?: CogniteInternalId[];
+  parentExternalIds?: CogniteExternallId[];
+  rootIds?: IdEither[];
+  source?: string;
+  root?: boolean;
+}
+
+export interface TimeseriesFilterRequestParams extends FilterRequestParams {
+  name?: string;
+  unit?: string;
+  isString?: boolean;
+  isStep?: boolean;
+  assetIds?: CogniteInternalId[];
+  assetExternalIds?: CogniteExternallId[];
+  rootAssetIds?: IdEither[];
+}
+
+export interface EventsFilterRequestParams extends FilterRequestParams {
+  startTime?: TimeRange;
+  endTime?: TimeRange;
+  assetIds?: CogniteInternalId[];
+  assetExternalIds?: CogniteExternallId[];
+  rootAssetIds?: IdEither[];
+  source?: string;
+  type?: string;
+  subtype?: string;
+}
+
+export interface FilterRequest<Filter> extends Limit, Cursor {
+  filter: Filter;
+}
+
+export interface TimeSeriesResponseItem {
+  id: number;
+  externalId?: string;
+  name?: string;
+  isString?: boolean;
+  metadata?: object;
+  unit?: string;
+  assetId?: string;
+  isStep: boolean;
+  description?: string;
+  source?: string;
+  sourceId?: string;
+  createdTime: number;
+  lastUpdatedTime: number;
+  selected: boolean;
+}
+
+export type TimeSeriesResponse = Items<TimeSeriesResponseItem>;
+
+export type Limit = {
+  limit?: number;
+};
+
+export interface Cursor {
+  cursor?: string;
+}
+
+
+export interface Metadata {
+  [name: string]: string;
+}
+
+export type TimeRange = Range<number>;
+export type CogniteInternalId = number;
+export type CogniteExternallId = string;
+
+
+
+export interface Timestamp {
+  timestamp: number;
+}
+
+export type Items<T = object> = {
+  items: T[];
+};
+
+export type IdEither =
+  | {
+      id: number;
+    }
+  | {
+      externalId: string;
+    };
+
+export interface Range<T> {
+  min?: T;
+  max?: T;
+}
+
